edit function not working, everything nearly finr

+++ projects/main.py
@@ -43,23 +43,39 @@ def profile():
     
     return render_template('/auth/profile.html', user=current_user, posts=posts, comments=comments)  # Render the profile.html template with the current user
 
+"""
+No need to use 'own' flag as we can use thread_author_id = current_user.id directly in different function.
+
+thread_detailed function: if users not login, just show forum.html(default) and thread_detailed (written in def thread_detailed() )
+"""
 @main.route('/thread_detailed/<int:thread_id>')  # Route decorator for the thread_detailed route
 def thread_detailed(thread_id):
     thread = db.get_or_404(Thread, thread_id)  # Get the thread with the specified thread_id from the database
     thread.count = db.session.query(Comment).filter(Comment.thread_id == thread.id).count()  # Count the number of comments for the thread
-    try:
-        if int(thread.author.id) == int(current_user.id):  # Check if the current user is the author of the thread
-            return render_template("thread_detailed.html", thread=thread, user=current_user, edit=False, own=True)  # Render the thread_detailed.html template with the thread, current user, and edit and own flags
-        else:
-            return render_template("thread_detailed.html", thread=thread, user=current_user, edit=False, own=False)  # Render the thread_detailed.html template with the thread, current user, and edit and own flags
-    except:
-        return render_template("thread_detailed.html", thread=thread, user=current_user, edit=False, own=False)  # Render the thread_detailed.html template with the thread, current user, and edit and own flags
+    # try:
+    #     if int(thread.author.id) == int(current_user.id):  # Check if the current user is the author of the thread
+    #         return render_template("thread_detailed.html", thread=thread, user=current_user, edit=False, own=True)  # Render the thread_detailed.html template with the thread, current user, and edit and own flags
+    #     else:
+    #         return render_template("thread_detailed.html", thread=thread, user=current_user, edit=False, own=False)  # Render the thread_detailed.html template with the thread, current user, and edit and own flags
+    # except:
+    #     return render_template("thread_detailed.html", thread=thread, user=current_user, edit=False, own=False)  # Render the thread_detailed.html template with the thread, current user, and edit and own flags
+    if current_user.is_authenticated:
+        # if using login account, allow to have ability to edit thread and comment (Just ability, specifically when they can use will be on different function)
+        return render_template("thread_detailed.html", thread=thread, edit=False, edit_comment_id = None)
+    else:
+        # if not login, just be able to access to thread_detailed.html
+        return render_template("thread_detailed.html", thread=thread)
 
 @main.route('/thread_detailed/edit/<int:thread_id>')  # Route decorator for the thread_edit route
 def thread_edit(thread_id):
     thread = db.get_or_404(Thread, thread_id)  # Get the thread with the specified thread_id from the database
-    thread.count = db.session.query(Comment).filter(Comment.thread_id == thread.id).count()  # Count the number of comments for the thread
-    return render_template("thread_detailed.html", thread=thread, user=current_user, edit=True)  # Render the thread_detailed.html template with the thread, current user, and edit flag
+    if current_user.id == thread.author.id:
+        # if users own that thread, allow edit
+        thread.count = db.session.query(Comment).filter(Comment.thread_id == thread.id).count()  # Count the number of comments for the thread
+        return render_template("thread_detailed.html", thread=thread, user=current_user, edit=True, edit_comment_id=None)  # Render the thread_detailed.html template with the thread, current user, and edit flag
+    else:
+        # if not own thread, just allow to see as non-login user, redirect to thread_detailed.html
+        return redirect(url_for("main.thread_detailed", thread_id=thread.id))
 
 @main.route('/add')  # Route decorator for the add_page route
 def add_page():
@@ -87,18 +103,24 @@ def add_thread():
 @main.route("/thread_detailed/delete/<int:thread_id>", methods=["POST"])  # Route decorator for the del_thread route
 def del_thread(thread_id):
     thread = db.get_or_404(Thread, thread_id)  # Get the thread with the specified thread_id from the database
-    db.session.delete(thread)  # Delete the thread from the database
-    db.session.commit()  # Commit the changes to the database
+    #Only allow delete for users who own thread and Admin
+    if current_user.id == thread.author.id or current_user.id == 1:
+        db.session.delete(thread)  # Delete the thread from the database
+        db.session.commit()  # Commit the changes to the database
     return redirect(url_for('main.index'))  # Redirect to the index route
 
 @main.route("/thread_detailed/editing/<int:thread_id>", methods=["POST"])  # Route decorator for the thread_update route
 def thread_update(thread_id):
     thread = db.get_or_404(Thread, thread_id)  # Get the thread with the specified thread_id from the database
-    if request.form["title"] == "":  # Check if the title field is not empty
-        return redirect(url_for('main.thread_detailed', thread_id=thread.id))  # Redirect to the thread_detailed route
-    thread.title = request.form["title"]  # Update the title and content of the thread
-    thread.content = request.form["content"]
-    db.session.commit()  # Commit the changes to the database
+    # only allow own users to edit their own thread
+    if current_user.id == thread.author.id:
+        if request.form["title"] == "":  # Check if the title field is not empty
+            # (FLASH HERE)
+            return redirect(url_for('main.thread_detailed', thread_id=thread.id))  # Redirect to the thread_detailed route
+        elif request.form["title"] != "":
+            thread.title = request.form["title"]  # Update the title and content of the thread
+            thread.content = request.form["content"]
+            db.session.commit()  # Commit the changes to the database
     return redirect(url_for('main.thread_detailed', thread_id=thread.id))  # Redirect to the thread_detailed route
 
 @main.route("/thread_detailed/<int:thread_id>", methods=["POST"])  # Route decorator for the add_comment route
@@ -122,9 +144,10 @@ def del_subheading(subheading_id):
 
 @main.route('/del/comment/<int:comment_id>', methods=["POST"])  # Route decorator for the del_comment route
 def del_comment(comment_id):
-    comment = db.get_or_404(Comment, comment_id)  # Get the comment with the specified comment_id from the database
-    db.session.delete(comment)  # Delete the comment from the database
-    db.session.commit()  # Commit the changes to the database
+    comment = db.get_or_404(Comment, comment_id) # Get the comment with the specified comment_id from the database
+    if current_user.id == comment.author.id or current_user.id == 1:
+        db.session.delete(comment)  # Delete the comment from the database
+        db.session.commit()  # Commit the changes to the database
     return redirect(url_for('main.thread_detailed', thread_id=comment.thread_id))  # Redirect to the thread_detailed route
 
 @main.route('/update<int:id>', methods=['GET', 'POST'])  # Route decorator for the update route
@@ -194,24 +217,6 @@ def upvote():
     else:
         return jsonify({'status': 'error', 'message': 'No thread_id provided'}), 400  # Return error status if no thread_id is provided
     
-# Route for reply comment
-# @main.route('/reply_comment/<int:comment_id>', methods=['POST'])
-# @login_required
-# def reply_comment(comment_id):
-#     parent_comment = Comment.query.get_or_404(comment_id)
-#     content = request.form.get('content')
-#     reply = Comment(content=content, parent=parent_comment, author=current_user, thread_id=parent_comment.thread_id)
-#     db.session.add(reply)
-#     db.session.commit()
-#     return redirect(url_for('main.thread_detailed', thread_id=parent_comment.thread.id))
-
-# # Route for edit comment
-# @main.route('/edit_comment/<int:comment_id>')
-# @login_required
-# def edit_comment(comment_id):
-#     comment = Comment.query.get_or_404(comment_id)
-#     return render_template('thread_detailed.html', thread=comment.thread, user=current_user, edit_comment=True, comment_id=comment_id)
-
 @main.route("/comment_reply/<int:comment_id>", methods=["POST"])  # Route decorator for the add_reply route
 def add_reply(comment_id):
     if request.form["content"] != "":  # Check if the content field is not empty
@@ -240,18 +245,24 @@ def comment_detailed(comment_id):
 @main.route('/comment_detailed/edit/<int:comment_id>')  # Route decorator for the comment_edit route
 def comment_edit(comment_id):
     comment = db.get_or_404(Comment, comment_id)  # Get the comment with the specified comment_id from the database
-    comment.count = db.session.query(Comment).filter(Comment.parent_id == comment.id).count()  # Count the number of replies for the comment
-    thread = comment.thread  # Get the thread that the comment belongs to
-    thread.count = db.session.query(Comment).filter(Comment.thread_id == thread.id).count()  # Count the number of comments for the thread
-    return render_template("thread_detailed.html", thread=thread, comment=comment, user=current_user, edit_comment_id=comment_id)  # Render the thread_detailed.html template with the thread, comment, current user, and edit flag
+    thread = db.get_or_404(Thread, comment.thread_id)
+    if current_user.id == comment.author.id:
+        # comment.count = db.session.query(Comment).filter(Comment.parent_id == comment.id).count()  # Count the number of replies for the comment
+        thread.count = db.session.query(Comment).filter(Comment.thread_id == thread.id).count()  # Count the number of comments for the thread
+        return render_template("thread_detailed.html", thread=thread, user=current_user, edit=False, edit_comment_id=comment_id)  
+    else:
+        return redirect(url_for('main.thread_detailed', thread_id = comment.thread.id))
 
 @main.route("/comment_detailed/editing/<int:comment_id>", methods=["POST"])  # Route decorator for the comment_update route
 def comment_update(comment_id):
     comment = db.get_or_404(Comment, comment_id)  # Get the comment with the specified comment_id from the database
-    if request.form["content"] == "":  # Check if the content field is not empty
-        return redirect(url_for('main.thread_detailed', thread_id=comment.thread.id))  # Redirect to the thread_detailed route
-    comment.content = request.form["content"]  # Update the content of the comment
-    db.session.commit()  # Commit the changes to the database
+    if current_user.id == comment.author.id:
+        if request.form["content"] == "":  # Check if the content field is not empty
+            # FLASH HERE
+            return redirect(url_for('main.thread_detailed', thread_id=comment.thread.id))  # Redirect to the thread_detailed route
+        elif request.form['content'] != "":
+            comment.content = request.form["content"]  # Update the content of the comment
+        db.session.commit()  # Commit the changes to the database
     return redirect(url_for('main.thread_detailed', thread_id=comment.thread.id))  # Redirect to the thread_detailed route
